refactor: 프로퍼티 관련 클래스 패키지 분리
- 프로퍼티 관련 클래스만 따로 분리합니다.
