+++ src/main/java/dev/bang/pickcar/config/SwaggerConfig.java
@@ -1,5 +1,6 @@
 package dev.bang.pickcar.config;
 
+import dev.bang.pickcar.config.properties.SwaggerProperties;
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
