+++ src/main/java/dev/bang/pickcar/config/properties/SwaggerProperties.java
@@ -1,4 +1,4 @@
-package dev.bang.pickcar.config;
+package dev.bang.pickcar.config.properties;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
